pyproject.toml
[project]
name = "englishbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.26.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.5",
]
